+++ Runtime/Controllers/AbstractController.cs
@@ -37,9 +37,14 @@ namespace ActionCode.UISystem
         protected virtual void OnEnable() => Reload();
         protected virtual void OnDisable() => UnsubscribeEvents();
 
-        public virtual void Focus() => Root.Focus();
         public virtual void Activate() => gameObject.SetActive(true);
         public virtual void Deactivate() => gameObject.SetActive(false);
+
+        public virtual async Awaitable FocusAsync()
+        {
+            Root.Focus();
+            await Awaitable.NextFrameAsync();
+        }
 
         public bool IsValid() => Root != null;
         public void Show() => SetVisibility(true);