+++ CHANGELOG.md
@@ -6,10 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- MenuController
 - LocalizedString
 - Confirmation Popup
 - Dialogue Popup
 - UITK simple animators
+- ActionCode.ScreenFadeSystem package dependency
 
 ## [0.2.0] - 2025-05-04
 ### Added